requirements.txt
streamlit
langchain
langchain-community
langchain-experimental
transformers
pandas
python-dotenv 